Remove example in libraries and adjust variable type

+++ Arduino/src/Function.h
@@ -20,7 +20,7 @@
 #define CONTINU true
 
 void clampValue(int8_t *_val, uint8_t _min, uint8_t _max);
-bool timer(unsigned long currentTime, unsigned long *oldTime, unsigned long delay);
+bool timer(uint32_t currentTime, uint32_t *oldTime, uint16_t delay);
 bool buffercmp(char originArray[], char targetArray[], uint16_t BufferLength);
 void bufferCopy(char originArray[], char targetArray[], int8_t index, uint8_t wordSize);
 uint8_t motionSense(int8_t index, int8_t lastIndex);
